refactor(config): fix typo in environments interfaces import path
Update import path from 'environtments.interfaces' to 'environments.interfaces' to match the correct file name
