+++ src/domain/shared/interfaces/environments.interfaces.ts
